Put hash storage back in block storage
